refactor: uint256 present value
- cleaner missed interval getter

+++ src/MToken.sol
@@ -198,7 +198,7 @@ contract MToken is IMToken, ContinuousIndexing, ERC20Extended {
         emit Transfer(account_, address(0), amount_);
 
         if (_balances[account_].isEarning) {
-            // NOTE: When burning a present amount, round the principal up in favor of the Minter Gateway.
+            // NOTE: When burning a present amount, round the principal up in favor of the protocol.
             _subtractEarningAmount(account_, _getPrincipalAmountRoundedUp(UIntMath.safe128(amount_)));
             updateIndex();
         } else {
@@ -215,7 +215,7 @@ contract MToken is IMToken, ContinuousIndexing, ERC20Extended {
         emit Transfer(address(0), recipient_, amount_);
 
         if (_balances[recipient_].isEarning) {
-            // NOTE: When minting a present amount, round the principal down in favor of the Minter Gateway.
+            // NOTE: When minting a present amount, round the principal down in favor of the protocol.
             _addEarningAmount(recipient_, _getPrincipalAmountRoundedDown(UIntMath.safe128(amount_)));
             updateIndex();
         } else {
@@ -324,7 +324,7 @@ contract MToken is IMToken, ContinuousIndexing, ERC20Extended {
 
         // If this is an in-kind transfer, then...
         if (senderIsEarning_ == _balances[recipient_].isEarning) {
-            // NOTE: When subtracting a present amount from an earner, round the principal up in favor of the Minter Gateway.
+            // NOTE: When subtracting a present amount from an earner, round the principal up in favor of the protocol.
             return
                 _transferAmountInKind( // perform an in-kind transfer with...
                     sender_,
@@ -336,12 +336,12 @@ contract MToken is IMToken, ContinuousIndexing, ERC20Extended {
         // If this is not an in-kind transfer, then...
         if (senderIsEarning_) {
             // either the sender is earning and the recipient is not, or...
-            // NOTE: When subtracting a present amount from an earner, round the principal up in favor of the Minter Gateway.
+            // NOTE: When subtracting a present amount from an earner, round the principal up in favor of the protocol.
             _subtractEarningAmount(sender_, _getPrincipalAmountRoundedUp(safeAmount_));
             _addNonEarningAmount(recipient_, safeAmount_);
         } else {
             // the sender is not earning and the recipient is.
-            // NOTE: When adding a present amount to an earner, round the principal down in favor of the Minter Gateway.
+            // NOTE: When adding a present amount to an earner, round the principal down in favor of the protocol.
             _subtractNonEarningAmount(sender_, safeAmount_);
             _addEarningAmount(recipient_, _getPrincipalAmountRoundedDown(safeAmount_));
         }
@@ -369,7 +369,7 @@ contract MToken is IMToken, ContinuousIndexing, ERC20Extended {
 
     /**
      * @dev   Returns the present amount (rounded down) given the principal amount, using the current index.
-     *        All present amounts are rounded down in favor of the Minter Gateway.
+     *        All present amounts are rounded down in favor of the protocol.
      * @param principalAmount_ The principal amount.
      */
     function _getPresentAmount(uint128 principalAmount_) internal view returns (uint128 amount_) {
@@ -378,12 +378,12 @@ contract MToken is IMToken, ContinuousIndexing, ERC20Extended {
 
     /**
      * @dev   Returns the present amount (rounded down) given the principal amount and an index.
-     *        All present amounts are rounded down in favor of the Minter Gateway, since they are assets.
+     *        All present amounts are rounded down in favor of the protocol, since they are assets.
      * @param principalAmount_ The principal amount.
      * @param index_           An index
      */
     function _getPresentAmount(uint128 principalAmount_, uint128 index_) internal pure returns (uint128 amount_) {
-        return _getPresentAmountRoundedDown(principalAmount_, index_);
+        return uint128(_getPresentAmountRoundedDown(principalAmount_, index_)); // TODO: fix cast.
     }
 
     /**
